Version 4.0.2

+++ src/models/ReviewStatsModel.php
@@ -32,9 +32,9 @@ class ReviewStatsModel extends Model
     // =========================================================================
 
     /**
-     * @var int|null Average Rating
+     * @var float|null Average Rating
      */
-    public ?int $averageRating;
+    public ?float $averageRating;
 
     /**
      * @var int|null Total Ratings
